+++ src/main/java/com/example/searchengine/repository/SiteRepository.java
@@ -3,12 +3,12 @@ package com.example.searchengine.repository;
 import com.example.searchengine.models.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.example.searchengine.models.Site;
+import com.example.searchengine.config.Site;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface SiteRepository extends JpaRepository<Site, Long> {
+public interface SiteRepository extends JpaRepository<Site, Integer> {
 
     List<Site> findByUrlContaining(String url);
 
@@ -16,7 +16,7 @@ public interface SiteRepository extends JpaRepository<Site, Long> {
 
     Optional<Site> findByUrl(String url);
 
-    List<Long> findByGroupId(Long groupId);
+    List<Integer> findByGroupId(Integer groupId);
 
     Site findByDomain(String domain);
